+++ MauiProgram.cs
@@ -1,4 +1,4 @@
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Uninstaller.Helpers;
 using Uninstaller.Services;
 
@@ -28,6 +28,9 @@ public static class MauiProgram
 #if ANDROID
         builder.Services.AddSingleton<IAppInventoryService, Platforms.Android.AppInventoryService>();
         builder.Services.AddSingleton<IToastService, Platforms.Android.ToastService>();
+#elif WINDOWS
+        builder.Services.AddSingleton<IAppInventoryService, Platforms.Windows.AppInventoryService>();
+        builder.Services.AddSingleton<IToastService, Platforms.Windows.ToastService>();
 #endif
 
         // Las paginas (carpeta Pages) las instancia el Shell por DataTemplate y resuelven sus